cdefs.h: rename from kilipili_cdefs.h
