+++ Devices/BlockDeviceFile.cpp
@@ -3,7 +3,7 @@
 // https://opensource.org/licenses/BSD-2-Clause
 
 #include "BlockDeviceFile.h"
-#include "kilipili_cdefs.h"
+#include "cdefs.h"
 #include <memory>
 
 namespace kio::Devices
