+++ Graphics/graphics_types.h
@@ -3,7 +3,7 @@
 // https://opensource.org/licenses/BSD-2-Clause
 
 #pragma once
-#include "kilipili_cdefs.h"
+#include "cdefs.h"
 #include "standard_types.h"
 
 /*	This file defines some enum types used in the Pixmap and graphics engine.
