+++ common/geometry.h
@@ -4,7 +4,7 @@
 
 #pragma once
 #include "basic_math.h"
-#include "kilipili_cdefs.h"
+#include "cdefs.h"
 #include <utility>
