+++ common/sort.h
@@ -3,7 +3,7 @@
 // BSD-2-Clause license
 // https://opensource.org/licenses/BSD-2-Clause
 
-#include "kilipili_cdefs.h"
+#include "cdefs.h"
 #include "relational_operators.h"
 #include "template_helpers.h"
